+++ src/pages/admin/resource/ResourceManage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useState, useEffect } from "react";
 import { RightContainer, TitleText, WhiteContainer } from "components/rightContainer/RightContainer";
-import {Bar, BookedTable, BookedThead, NoLineTr, TableContainer} from "../../booking/bookedList/BookedList";
+import {Bar, BookedTable, BookedThead, NoLineTr, TableContainer} from "../../basic/myBookings/BookedList";
 import ResourceManageTableCell from "./ResourceManageTableCell";
 import ManageSearchBar from "components/searchBar/ManageSearchBar";
 import { getToken } from "utils/IsLoginUtil";
