+++ src/pages/admin/resource/ResourceManageDetail.js
@@ -1,48 +1,92 @@
 import React from 'react';
-import { OfficesAxios, AdminBookingOfficeAxios } from 'api/AxiosApi';
+import styled from "styled-components"
+import { ResourcesAxios, AdminBookingResourceAxios } from 'api/AxiosApi';
 import { useState, useEffect } from "react";
 import { useParams } from 'react-router-dom';
 import Capsule from 'components/capsule/Capsule';
 import { basicError } from 'utils/ErrorHandlerUtil';
-import { Bar } from 'pages/booking/bookedList/BookedList';
+import { Bar } from 'pages/basic/myBookings/BookedList';
 import { RightContainer, WhiteContainer,TitleText } from 'components/rightContainer/RightContainer';
 import { MainTextContainer, SubTextContainer, SelectedSubTitleText, UnselectedSubTitleText } from 'components/officeBooking/SubTitleBar';
+import ResourceInfo from 'components/resourceInfo/ResourceInfo';
 import { getToken } from 'utils/IsLoginUtil';
 import MoreButtonIcon from "../../../assets/images/button/triple_dot_icon.svg"
-import { MoreButton, OptionButton, OptionsView, InfoTable, InfoTableData } from '../resourceManage/ResourceManageDetail';
-import OfficeInfo from 'components/officeInfo/OfficeInfo';
 
-function OfficeManageDetail() {
-    let { officeId } = useParams()
+export const MoreButton = styled.button`
+    border: none;
+    background-color: #2A3042;
+    float: right;
+    margin: 15px; 10px; 5px; 0;
+`
+export const OptionsView = styled.div`
+    display: ${props => props.isShowing ? 'table-cell' : 'none'};
+    width: 85px;
+    height: 96px;
+    background: white;
+    border: 1px solid #545F71;
+    border-radius: 12px;
+    position: absolute;
+    top: 35px;
+    right: 30px;
+    vertical-align: middle;
+`
+
+export const OptionButton = styled.button`
+    width: 85px;
+    height: 48px;
+    background: none;
+    border: none;
+    color: ${props => props.isDelete ? '#A65959' : '#545F71'};
+    font-size: 16px;
+`
+
+export const InfoTable = styled.table`
+    width: 94%;
+    margin: 40px 0 40px 40px;
+    border-collapse : collapse;
+    border: 1px solid #959494;
+    font-size: 17px;
+    align: center;
+`
+
+export const InfoTableData = styled.td`
+    border: 1px solid #959494;
+    height: 45px
+`
+
+function ResourceManageDetail() {
+    let { resourceId } = useParams()
 
     const [isShowingOptions, setOptionViewShowing] = useState(false)
-    const [officeInfo, setOfficeInfo] = useState([])
+    const [resourceInfo, setResourceInfo] = useState([])
     const [bookingsInfo, setBookingsInfo] = useState([])
 
-    const getOfficeInfo = () => {
-        OfficesAxios.get(`/${officeId}`, {
-            headers: {
-                Authorization: getToken()
-            }
-        })
-        .then((Response)=>{ setOfficeInfo(Response.data.data) })
-        .catch((Error)=>{ 
-            basicError(Error) 
-            console.log(Error)
-            window.alert("회의실 정보를 불러올 수 없습니댜.")
-            window.history.back()
-        })
-    }
-
-    const getOfficeBookingListInfo = () => {
-        AdminBookingOfficeAxios.get(`offices/${officeId}`, {
+    const getResourceInfo = () => {
+        ResourcesAxios.get(`/${resourceId}`, {
             headers: {
                 Authorization: getToken()
             }
         })
         .then((Response)=>{ 
-            setBookingsInfo(Response.data.data.officesLists) 
+            console.log(Response.data.data)
+            setResourceInfo(Response.data.data) })
+        .catch((Error)=>{ 
+            basicError(Error) 
+            console.log(Error)
+            window.alert("장비 정보를 불러올 수 없습니댜.")
+            window.history.back()
         })
+    }
+
+    const getResourceBookingListInfo = () => {
+        AdminBookingResourceAxios.get(`${resourceId}`, {
+            headers: {
+                Authorization: getToken()
+            }
+        })
+        .then((Response)=>{ 
+            console.log(Response.data.data)
+            setBookingsInfo(Response.data.data.resourcesLists) })
         .catch((Error)=>{ 
             basicError(Error) 
             console.log(Error)
@@ -52,68 +96,66 @@ function OfficeManageDetail() {
     }
 
     const deleteResource = () => {
-        if (window.confirm("회의실을 삭제하시겠습니까?")) {
-            AdminBookingOfficeAxios.delete(`${officeId}`, {
+        if (window.confirm("장비을 삭제하시겠습니까?")) {
+            AdminBookingResourceAxios.delete(`${resourceId}`, {
                 headers: {
                     Authorization: getToken()
                 }
             })
-            .then((Response)=>{ 
+            .then((Response)=>{
                 if(Response.data.status === 200) {
-                    alert('회의실을 성공적으로 삭제하였습니다.')
+                    alert('장비을 성공적으로 삭제하였습니다.')
                     window.history.back()
-                } 
+                }
             })
-            .catch((Error)=>{ 
-                basicError(Error) 
+            .catch((Error)=>{
+                basicError(Error)
                 console.log(Error)
             })
             setOptionViewShowing(false)
-        } 
+        }
     }
 
     useEffect(() => {
-        getOfficeInfo()
-        getOfficeBookingListInfo()
+        getResourceInfo()
+        getResourceBookingListInfo()
     }, []);
 
     return <RightContainer>
-        <TitleText>회의실 관리</TitleText>
+        <TitleText>장 관리</TitleText>
 
         <WhiteContainer style={{display:'inline'}}>
             <Bar style={{position:'static'}}>
                 <MainTextContainer>
-                    <SelectedSubTitleText>{officeInfo.name}</SelectedSubTitleText>
+                    <SelectedSubTitleText>{resourceInfo.name}</SelectedSubTitleText>
                 </MainTextContainer>
                 <SubTextContainer>
-                    <UnselectedSubTitleText>{officeInfo.location}</UnselectedSubTitleText>
+                    <UnselectedSubTitleText>{resourceInfo.location}</UnselectedSubTitleText>
                 </SubTextContainer>
-                <MoreButton onClick={() => { setOptionViewShowing(!isShowingOptions) }}> 
+                <MoreButton onClick={() => { setOptionViewShowing(!isShowingOptions) }}>
                     <img src={MoreButtonIcon} alt="더보기" />
                 </MoreButton>
                 <OptionsView isShowing={isShowingOptions}>
-                    <OptionButton onClick={() => {  }}>수정</OptionButton>
+                    <OptionButton onClick={() => { window.location.href = `/admin/resources/edit/${resourceId}` }}>수정</OptionButton>
                     <OptionButton isDelete={true} onClick={ deleteResource }>삭제</OptionButton>
                 </OptionsView>
             </Bar>
 
-            <OfficeInfo isDetailPage={true}
-                    key={officeInfo.name}
-                    capacity={officeInfo.capacity}
-                    facilityList={officeInfo.facilityList}
-                    description={officeInfo.description}
-                    imgUrl={officeInfo.imgUrl}
-                />
+            <ResourceInfo detail={false}
+                        responsibilityName={resourceInfo.responsibilityName}
+                        responsibilityPhone={resourceInfo.responsibilityPhone}
+                        description={resourceInfo.description}
+                        imgUrl={resourceInfo.imgUrl} />
 
             <div style={{width:'150px', padding:'30px 0 0 40px'}}>
                 <Capsule color="purple" text="예약 내역"/>
             </div>
-            
+
             <InfoTable>
                 <tr style={{backgroundColor:'#D0B1EE', border: '1px solid #959494', height:'45px'}}>
                     <th scope="col" style={{width: '20%', border:'1px solid #959494'}} >요청자</th>
-                    <th scope="col" style={{width: '28%', border:'1px solid #959494'}} >예약일자</th>
-                    <th scope="col" style={{width: '40%', border:'1px solid #959494'}} >목적</th>
+                    <th scope="col" style={{width: '25%', border:'1px solid #959494'}} >예약일자</th>
+                    <th scope="col" style={{width: '43%', border:'1px solid #959494'}} >목적</th>
                     <th scope="col" style={{width: '12%', border:'1px solid #959494'}} >상태</th>
                 </tr>
                 {bookingsInfo.map(function(info){
@@ -131,4 +173,4 @@ function OfficeManageDetail() {
         </WhiteContainer>
     </RightContainer>
 }
-export default OfficeManageDetail;
+export default ResourceManageDetail;
