+++ src/pages/admin/resourceBookings/ResourceBookingManage.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useState, useEffect } from "react";
 import axios from "axios";
 import { RightContainer, TitleText, WhiteContainer } from "components/rightContainer/RightContainer";
-import { Bar, BookedTable, BookedThead, TableContainer } from "../../booking/bookedList/BookedList";
+import { Bar, BookedTable, BookedThead, TableContainer } from "../../basic/myBookings/BookedList";
 import ManageSearchBar from "components/searchBar/ManageSearchBar";
 import ResourceBookingManageCell from "./ResourceBookingManageCell";
 import { AdminBookingAxios } from "api/AxiosApi";
