+++ src/pages/admin/user/UserManageLine.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import {BookedLineTr} from 'pages/booking/bookedList/BookedList';
+import {BookedLineTr} from 'pages/basic/myBookings/BookedList';
 import {RoleCapsule} from 'components/capsule/RoleCapsule';
 import MoreIcon from 'assets/images/moreIcon.svg'
 import {UserMoreModal} from "./UserMoreModal";
